make metric sorting a hook + use it
