+++ src/pages/organize/[orgId]/projects/[projectId]/areaassignments/[areaAssId]/report.tsx
@@ -53,7 +53,7 @@ import {
 } from 'features/areaAssignments/types';
 import useAreaAssignmentMetrics from 'features/areaAssignments/hooks/useAreaAssignmentMetrics';
 import ZUILockCard from 'zui/ZUILockCard';
-import sortMetrics from 'features/canvass/utils/sortMetrics';
+import useSortedMetrics from 'features/canvass/hooks/useSortedMetrics';
 
 const scaffoldOptions = {
   authLevelRequired: 2,
@@ -88,7 +88,7 @@ const AreaAssignmentReportPage: PageWithLayout<AreaAssignmentReportProps> = ({
   const messages = useMessages(messagesIds);
   const { showConfirmDialog } = useContext(ZUIConfirmDialogContext);
   const metricsList = useAreaAssignmentMetrics(parseInt(orgId), areaAssId);
-  const metrics = sortMetrics(metricsList);
+  const metrics = useSortedMetrics(metricsList);
 
   const [metricBeingCreated, setMetricBeingCreated] =
     useState<Partial<ZetkinMetric> | null>(null);
